+++ com/familytree/data/Familytree_gedcom_yadav.ged
@@ -739,3 +739,130 @@
 
 
 0 NOTE ############################### US_02_06 End ###############################
+
+0 NOTE ############################### US_03_04 start ###############################
+
+0 NOTE Rahul /Kampati/
+0 NOTE three generations of Vadde family
+0 NOTE define Mahesh Vadde
+0 US0304I01 INDI
+1 NAME Mahesh /Vadde/ 
+1 BIRT
+2 DATE 04 APR 2006 
+1 SEX M 
+1 FAMS US0304F01 
+1 DEAT
+2 DATE 13 JUN 2005
+0 NOTE define Buvi Vadde
+0 US0304I02 INDI 
+1 NAME Buvi /Vadde/ 
+1 BIRT 
+2 DATE 14 MAR 1922 
+1 SEX F 
+1 FAMS US0304F01 
+0 NOTE define Mukesh Vadde
+0 US0304I03 INDI 
+1 NAME Mukesh /Vadde/ 
+1 BIRT 
+2 DATE 28 APR 1945 
+1 SEX M 
+1 FAMS US0304F02
+1 FAMC US0304F01
+0 NOTE define Kavya Sahu
+0 US0304I04 INDI
+1 NAME Kavya /Sahu/
+1 BIRT 
+2 DATE 15 DEC 1947
+1 SEX F
+1 FAMC US0304F01 
+1 FAMS US0304F03
+1 FAMS US0304F04
+0 NOTE define Madhu Vadde
+0 US0304I05 INDI
+0 NAME Madhu /Vadde/ 
+1 BIRT 
+2 DATE 14 MAR 1951
+1 SEX F
+1 FAMS US0304F02 
+0 NOTE define Haritha Vadde
+0 US0304I06 INDI
+0 NAME Haritha /Vadde/ 
+1 BIRT 
+2 DATE 19 FEB 1975
+1 SEX F
+1 FAMC US0304F02
+0 NOTE define Anusha Vadde
+0 US0304I07 INDI
+0 NAME Anusha /Vadde/ 
+1 BIRT 
+2 DATE 04 FEB 1975
+1 SEX F
+1 FAMC US0304F02
+0 NOTE define Abhi Yepuri
+0 US0304I08 INDI
+1 NAME Abhi /Yepuri/
+1 BIRT 
+2 DATE 18 APR 1959
+1 SEX M 
+1 FAMS US0304F03
+1 DEAT
+2 DATE 08 SEP 1982
+0 NOTE define Varun Yepuri
+0 US0304I09 INDI
+1 NAME Varun /Yepuri/
+1 BIRT 
+2 DATE 12 OCT 1975
+1 SEX M 
+1 FAMC US0304F03
+0 NOTE define Vishu Vadde
+0 US0304I10 INDI
+1 NAME Vishu /Vadde/
+1 BIRT 
+2 DATE 19 DEC 1950
+1 SEX M 
+1 FAMS US0304F04
+0 NOTE define Harsha Sahu
+0 US0304I11 INDI
+1 NAME Harsha /Sahu/
+1 BIRT 
+2 DATE 05 MAR 1987
+1 SEX M 
+1 FAMC US0304F04
+0 NOTE define Moni Sahu
+0 US0304I12 INDI
+1 NAME Moni /Sahu/
+1 BIRT 
+2 DATE 05 APR 1989
+1 SEX F 
+1 FAMC US0304F04
+0 US0304F01 FAM 
+1 MARR 
+2 DATE 18 OCT 1942
+1 HUSB US0304I01 
+1 WIFE US0304I02
+1 CHIL US0304I03 
+1 CHIL US0304I04
+0 US0304F02 FAM 
+1 MARR 
+2 DATE 15 APR 1975
+1 HUSB US0304I03 
+1 WIFE US0304I05
+1 CHIL US0304I06 
+1 CHIL US0304I07
+0 US0304F03 FAM 
+1 MARR 
+2 DATE 07 AUG 1975
+1 HUSB US0304I08 
+1 WIFE US0304I04
+1 CHIL US0304I09 
+0 US0304F04 FAM
+1 MARR
+2 DATE 12 OCT 1985
+1 HUSB US0304I10 
+1 WIFE US0304I04
+1 CHIL US0304I11 
+1 CHIL US0304I12
+1 DIV
+2 DATE 14 OCT 1983
+
+0 NOTE ############################### US_03_04 end ###############################
